Time: 3 ms (94.22%), Space: 72.9 MB (60.07%) - LeetHub

+++ 1803-average-waiting-time/1803-average-waiting-time.java
@@ -4,8 +4,8 @@ class Solution {
         long ft = 0;
         long sum = 0;
         for(int i = 0;i < n;i++) {
-                long at = (long)customers[i][0];
-                long bt = (long)customers[i][1];
+                long at = customers[i][0];
+                long bt = customers[i][1];
                 if(at>ft) {
                     ft = at + bt;
                 } else {
